refactor(db): extract command locks + conflict checks to src/db/commandLocks.ts
Move all named-lock helpers, string normalisation utilities, SQL exists-check
builders, error classes (CommandNotFoundError, CommandConflictError), conflict
assertion functions, and runSerializedCommandWrite into their own module.

Also simplifies acquireNamedLock: the four chained if/throw branches are
replaced by an early-return on success followed by a single message lookup,
reducing cyclomatic complexity from 5 to 3.
